+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 [Unreleased](https://github.com/jshwi/docsig/compare/v0.36.1...HEAD)
 ------------------------------------------------------------------------
+### Added
+- add user friendly errors for the commandline
 
 [0.36.1](https://github.com/jshwi/docsig/releases/tag/v0.36.1) - 2023-11-17
 ------------------------------------------------------------------------