+++ CHANGELOG.md
@@ -5,7 +5,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-[Unreleased](https://github.com/jshwi/docsig/compare/v0.36.1...HEAD)
+[Unreleased](https://github.com/jshwi/docsig/compare/v0.37.0...HEAD)
+------------------------------------------------------------------------
+
+[0.37.0](https://github.com/jshwi/docsig/releases/tag/v0.37.0) - 2023-11-17
 ------------------------------------------------------------------------
 ### Added
 - raise error if file not found